+++ src/pages/projContracts/api/docusign/downloadContract.ts
@@ -1,5 +1,5 @@
-import { KintoneClient } from '../../../../../api/kintone';
-import { yumecocoDocusign } from '../../../../../config/settings';
+import { KintoneClient } from '../../../../api/kintone';
+import { yumecocoDocusign } from '../../../../config/settings';
 import { TypeOfForm } from '../../form';
 
 type DownloadResponse = {
