+++ src/pages/projContracts/api/docusign/getSenderViewUrl.test.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { yumecocoDocusign } from '../../../../../config/settings';
+import { yumecocoDocusign } from '../../../../config/settings';
 
 describe('sendContract', () => {
   it('should send contract', async () => {
