+++ src/pages/projContracts/parts/ContractContainer.tsx
