+++ src/pages/projContracts/parts/PreviewToolBar/DownloadContract.tsx
@@ -1,6 +1,6 @@
 import LoadingButton from '@mui/lab/LoadingButton';
 import { Tooltip } from '@mui/material';
-import { yumecocoDocusign } from '../../../../../config/settings';
+import { yumecocoDocusign } from '../../../../config/settings';
 import { RiFileExcel2Fill } from '@react-icons/all-files/ri/RiFileExcel2Fill';
