+++ src/pages/projContracts/parts/PreviewToolBar/RefreshButton.tsx
@@ -1,7 +1,7 @@
 import { IconButton, CircularProgress, Grow, Tooltip } from '@mui/material';
 import RefreshIcon from '@mui/icons-material/Refresh';
 import { useFormikContext } from 'formik';
-import { useSnackBar } from '../../../../../hooks';
+import { useSnackBar } from '../../../../hooks';
 import { useState } from 'react';
 import { getFormDataById } from '../../api/fetchRecord';
 import { TypeOfForm } from '../../form';
