+++ src/pages/projContracts/parts/PreviewToolBar/SendContract.tsx
@@ -1,13 +1,13 @@
 import { sendContract } from '../../api/docusign/sendContract';
 import LoadingButton from '@mui/lab/LoadingButton';
 import SendIcon from '@mui/icons-material/Send';
-import { useConfirmDialog, useSnackBar } from '../../../../../hooks';
+import { useConfirmDialog, useSnackBar } from '../../../../hooks';
 import { Button, Tooltip } from '@mui/material';
 import { useFormikContext } from 'formik';
 import { TypeOfForm } from '../../form';
-import { useBackdrop } from '../../../../../hooks/useBackdrop';
+import { useBackdrop } from '../../../../hooks/useBackdrop';
 import { useNavigate } from 'react-router-dom';
-import { pages } from '../../../../Router';
+import { pages } from '../../../Router';
