+++ src/pages/projContracts/parts/ProspectShortcuts.tsx
@@ -1,8 +1,8 @@
 import { useFormikContext } from 'formik';
 import { useNavigate } from 'react-router-dom';
 import { TypeOfForm } from '..';
-import { Shortcuts } from '../../../../components/ui/speedDials/Shortcuts';
-import { pages } from '../../../Router';
+import { Shortcuts } from '../../../components/ui/speedDials/Shortcuts';
+import { pages } from '../../Router';
 
 export const ProspectShortcuts = () => {
   const { values: {
