+++ src/pages/projContracts/parts/SearchProjField.tsx
@@ -1,9 +1,9 @@
 import { Autocomplete, TextField, Stack, Tooltip } from '@mui/material';
 import { useField } from 'formik';
 import { useEffect, useState } from 'react';
-import { useLazyEffect } from '../../../../hooks';
+import { useLazyEffect } from '../../../hooks';
 import { searchProjects } from '../api';
-import { Caption } from '../../../../components/ui/typographies';
+import { Caption } from '../../../components/ui/typographies';
 
 type Opt = {
   id: string,
